Multimedia player: Add hidden volume slider label
Doesn't function like a typical label... aria-describedby is being used to programmatically-associate it to the volume slider.

The volume slider was already using an aria-labelledby attribute to programmatically-associate it to the video's title - just like all the other controls. That causes the video's title to function as every control's label. So for each control, screen readers announce the video's title as a prefix, then the control's attributes, then the control's name (i.e. "Play", "Mute", etc...).

It would've been possible to add the label's ID at the end of the volume slider's aria-labelledby attribute, but that caused "Volume" to be announced before the slider control's attributes in NVDA + Firefox/Chrome. So it produced "Looking for a Job Volume 1 slider 100" instead of "Looking for a Job slider 100 Volume 1". Not ideal.

By comparison, aria-describedby gets announced after the volume slider's attributes and supersedes the title attribute. That relegates the label to secondary information, but is better overall since it's consistent with how all the other controls are announced.

Fixes #8939.